fix: some more housekeeping

+++ plugins/modules/vault_database_connection_delete.py
@@ -12,12 +12,12 @@ module: vault_database_connection_delete
 version_added: 6.2.0
 author:
   - Martin Chmielewski (@M4rt1nCh)
-short_description: Delete a Database Connection, identified by its O(connection_name), in a given path
+short_description: Delete a Database Connection
 requirements:
   - C(hvac) (L(Python library,https://hvac.readthedocs.io/en/stable/overview.html))
   - For detailed requirements, see R(the collection requirements page,ansible_collections.community.hashi_vault.docsite.user_guide.requirements).
 description:
-  - Deletes a L(Database Connection,https://hvac.readthedocs.io/en/stable/usage/secrets_engines/database.html#delete-connection) from HashiCorp Vault
+  - Deletes a L(Database Connection,https://hvac.readthedocs.io/en/stable/usage/secrets_engines/database.html#delete-connection), identified by its O(connection_name) from HashiCorp Vault
 notes:
   - This module always reports C(changed) status because it cannot guarantee idempotence.
   - Use C(changed_when) to control that in cases where the operation is known to not change state.
@@ -25,7 +25,7 @@ attributes:
   check_mode:
     support: partial
     details:
-      - In check mode, an empty response will be returned and the write will not be performed.
+      - In check mode, an empty response will be returned and the deletion will not be performed.
 extends_documentation_fragment:
   - community.hashi_vault.attributes
   - community.hashi_vault.attributes.action_group
@@ -38,6 +38,7 @@ options:
     type: str
     required: True
   engine_mount_point:
+    default: database
     description:
       - Specify the mount point used by the database engine.
       - Defaults to the default used by C(hvac).
@@ -47,10 +48,10 @@ EXAMPLES = r"""
 - name: Delete a Database Connection with the default mount point
   community.hashi_vault.vault_database_connection_delete:
     url: https://vault:8201
-    connection_name: SomeName
     auth_method: userpass
-    username: user
+    username: '{{ user }}'
     password: '{{ passwd }}'
+    connection_name: SomeName
   register: result
 
 - name: Display the result of the operation
@@ -60,11 +61,11 @@ EXAMPLES = r"""
 - name: Delete a Database Connection with a custom mount point
   community.hashi_vault.vault_database_connection_delete:
     url: https://vault:8201
+    auth_method: userpass
+    username: '{{ user }}'
+    password: '{{ passwd }}'
     engine_mount_point: db1
     connection_name: SomeName
-    auth_method: userpass
-    username: user
-    password: '{{ passwd }}'
   register: result
 
 - name: Display the result of the operation
@@ -119,44 +120,49 @@ def run_module():
             exception=HVAC_IMPORT_ERROR
         )
 
-    parameters = {}
-    engine_mount_point = module.params.get('engine_mount_point', None)
-    if engine_mount_point is not None:
-        parameters['mount_point'] = engine_mount_point
-    parameters["connection_name"] = module.params.get('connection_name')
+    if module.check_mode == False:
+      parameters = {}
+      engine_mount_point = module.params.get('engine_mount_point', None)
+      if engine_mount_point is not None:
+          parameters['mount_point'] = engine_mount_point
+      parameters["connection_name"] = module.params.get('connection_name')
 
-    module.connection_options.process_connection_options()
-    client_args = module.connection_options.get_hvac_connection_options()
-    client = module.helper.get_vault_client(**client_args)
+      module.connection_options.process_connection_options()
+      client_args = module.connection_options.get_hvac_connection_options()
+      client = module.helper.get_vault_client(**client_args)
 
-    try:
-        module.authenticator.validate()
-        module.authenticator.authenticate(client)
-    except (NotImplementedError, HashiVaultValueError) as e:
-        module.fail_json(msg=to_native(e), exception=traceback.format_exc())
+      try:
+          module.authenticator.validate()
+          module.authenticator.authenticate(client)
+      except (NotImplementedError, HashiVaultValueError) as e:
+          module.fail_json(msg=to_native(e), exception=traceback.format_exc())
 
-    try:
-        raw = client.secrets.database.delete_connection(**parameters)
-    except hvac.exceptions.Forbidden as e:
-        module.fail_json(msg="Forbidden: Permission Denied to path ['%s']." % engine_mount_point, exception=traceback.format_exc())
-    except hvac.exceptions.InvalidPath as e:
-        module.fail_json(
-            msg="Invalid or missing path ['%s']. Check the path." % (engine_mount_point),
-            exception=traceback.format_exc()
-        )
+      try:
+          raw = client.secrets.database.delete_connection(**parameters)
+      except hvac.exceptions.Forbidden as e:
+          module.fail_json(msg="Forbidden: Permission Denied to path ['%s']." % engine_mount_point, exception=traceback.format_exc())
+      except hvac.exceptions.InvalidPath as e:
+          module.fail_json(
+              msg="Invalid or missing path ['%s']. Check the path." % (engine_mount_point),
+              exception=traceback.format_exc()
+          )
 
-    if raw.status_code not in [200, 204]:
-        module.fail_json(
-            status='failure',
-            msg="Failed to delete database connection. Status code: %s" % raw.status_code,
-        )
+      if raw.status_code not in [200, 204]:
+          module.fail_json(
+              status='failure',
+              msg="Failed to delete database connection. Status code: %s" % raw.status_code,
+          )
+      module.exit_json(
+          data={
+              'status': 'success',
+              'status_code': raw.status_code,
+              'ok': raw.ok,
+          },
+          changed=True
+      )
+    
     module.exit_json(
-        data={
-            'status': 'success',
-            'status_code': raw.status_code,
-            'ok': raw.ok,
-        },
-        changed=True
+          data={}
     )
