+++ plugins/modules/vault_database_role_delete.py
@@ -20,7 +20,6 @@ description:
   - L(Delete a role definition,https://hvac.readthedocs.io/en/stable/usage/secrets_engines/database.html#delete-a-role)
 notes:
   - Applies to both static and dynamic roles
-  - The I(data) option is not treated as secret and may be logged. Use the C(no_log) keyword if I(data) contains sensitive values.
   - This module always reports C(changed) status because it cannot guarantee idempotence.
   - Use C(changed_when) to control that in cases where the operation is known to not change state.
 attributes:
@@ -36,6 +35,7 @@ extends_documentation_fragment:
   - community.hashi_vault.engine_mount
 options:
   engine_mount_point:
+    default: database
     description:
       - Specify the mount point used by the database engine.
       - Defaults to the default used by C(hvac).
@@ -48,6 +48,10 @@ options:
 EXAMPLES = r"""
 - name: Delete a Role with the default mount point
   community.hashi_vault.vault_database_role_delete:
+    url: https://vault:8201
+    auth_method: userpass
+    username: '{{ user }}'
+    password: '{{ passwd }}'
     role_name: SomeRole
   register: result
 
@@ -57,6 +61,10 @@ EXAMPLES = r"""
 
 - name: Delete a Role with a custom mount point
   community.hashi_vault.vault_database_role_delete:
+    url: https://vault:8201
+    auth_method: userpass
+    username: '{{ user }}'
+    password: '{{ passwd }}'
     engine_mount_path: db1
     role_name: SomeRole
   register: result
@@ -113,44 +121,49 @@ def run_module():
             exception=HVAC_IMPORT_ERROR
         )
 
-    parameters = {}
-    engine_mount_point = module.params.get('engine_mount_point', None)
-    if engine_mount_point is not None:
-        parameters['mount_point'] = engine_mount_point
-    parameters["role_name"] = module.params.get('role_name')
+    if module.check_mode == False:
+      parameters = {}
+      engine_mount_point = module.params.get('engine_mount_point', None)
+      if engine_mount_point is not None:
+          parameters['mount_point'] = engine_mount_point
+      parameters["role_name"] = module.params.get('role_name')
 
-    module.connection_options.process_connection_options()
-    client_args = module.connection_options.get_hvac_connection_options()
-    client = module.helper.get_vault_client(**client_args)
+      module.connection_options.process_connection_options()
+      client_args = module.connection_options.get_hvac_connection_options()
+      client = module.helper.get_vault_client(**client_args)
 
-    try:
-        module.authenticator.validate()
-        module.authenticator.authenticate(client)
-    except (NotImplementedError, HashiVaultValueError) as e:
-        module.fail_json(msg=to_native(e), exception=traceback.format_exc())
+      try:
+          module.authenticator.validate()
+          module.authenticator.authenticate(client)
+      except (NotImplementedError, HashiVaultValueError) as e:
+          module.fail_json(msg=to_native(e), exception=traceback.format_exc())
 
-    try:
-        raw = client.secrets.database.delete_role(**parameters)
-    except hvac.exceptions.Forbidden as e:
-        module.fail_json(msg="Forbidden: Permission Denied to path ['%s']." % engine_mount_point, exception=traceback.format_exc())
-    except hvac.exceptions.InvalidPath as e:
-        module.fail_json(
-            msg="Invalid or missing path ['%s']. Check the path." % (engine_mount_point),
-            exception=traceback.format_exc()
-        )
+      try:
+          raw = client.secrets.database.delete_role(**parameters)
+      except hvac.exceptions.Forbidden as e:
+          module.fail_json(msg="Forbidden: Permission Denied to path ['%s']." % engine_mount_point, exception=traceback.format_exc())
+      except hvac.exceptions.InvalidPath as e:
+          module.fail_json(
+              msg="Invalid or missing path ['%s']. Check the path." % (engine_mount_point),
+              exception=traceback.format_exc()
+          )
 
-    if raw.status_code not in [200, 204]:
-        module.fail_json(
-            status='failure',
-            msg="Failed to delete role. Status code: %s" % raw.status_code,
-        )
+      if raw.status_code not in [200, 204]:
+          module.fail_json(
+              status='failure',
+              msg="Failed to delete role. Status code: %s" % raw.status_code,
+          )
+      module.exit_json(
+          data={
+              'status': 'success',
+              'status_code': raw.status_code,
+              'ok': raw.ok,
+          },
+          changed=True
+      )
+
     module.exit_json(
-        data={
-            'status': 'success',
-            'status_code': raw.status_code,
-            'ok': raw.ok,
-        },
-        changed=True
+        data={}
     )
